fix(frontend): move graphql consts out of component files
