+++ frontend/relay-workflows-lib/lib/components/WorkflowRelay.tsx
@@ -1,4 +1,4 @@
-import { graphql, useLazyLoadQuery } from "react-relay";
+import { useLazyLoadQuery } from "react-relay";
 import { ResizableBox } from "react-resizable";
 import "react-resizable/css/styles.css";
 import { Box } from "@mui/material";
@@ -7,101 +7,10 @@ import { Visit, visitToText } from "@diamondlightsource/sci-react-ui";
 import type { WorkflowStatus } from "workflows-lib";
 import RetriggerWorkflow from "./RetriggerWorkflow";
 import React from "react";
-import { WorkflowRelayQuery as WorkflowRelayQueryType } from "./__generated__/WorkflowRelayQuery.graphql";
+import { workflowRelayQuery as WorkflowRelayQueryType } from "../graphql/__generated__/workflowRelayQuery.graphql";
 import { useParams, useNavigate } from "react-router-dom";
 import { useFetchedTasks, useSelectedTasks } from "./workflowRelayUtils";
-
-export const workflowRelayQuery = graphql`
-  query WorkflowRelayQuery($visit: VisitInput!, $name: String!) {
-    workflow(visit: $visit, name: $name) {
-      name
-      visit {
-        proposalCode
-        proposalNumber
-        number
-      }
-      templateRef
-      parameters
-      status {
-        __typename
-        ... on WorkflowPendingStatus {
-          message
-        }
-        ... on WorkflowRunningStatus {
-          startTime
-          message
-          tasks {
-            id
-            name
-            status
-            depends
-            dependencies
-            stepType
-            artifacts {
-              name
-              url
-              mimeType
-            }
-          }
-        }
-        ... on WorkflowSucceededStatus {
-          startTime
-          endTime
-          message
-          tasks {
-            id
-            name
-            status
-            depends
-            dependencies
-            stepType
-            artifacts {
-              name
-              url
-              mimeType
-            }
-          }
-        }
-        ... on WorkflowFailedStatus {
-          startTime
-          endTime
-          message
-          tasks {
-            id
-            name
-            status
-            depends
-            dependencies
-            stepType
-            artifacts {
-              name
-              url
-              mimeType
-            }
-          }
-        }
-        ... on WorkflowErroredStatus {
-          startTime
-          endTime
-          message
-          tasks {
-            id
-            name
-            status
-            depends
-            dependencies
-            stepType
-            artifacts {
-              name
-              url
-              mimeType
-            }
-          }
-        }
-      }
-    }
-  }
-`;
+import { workflowRelayQuery } from "../graphql/workflowRelayQuery";
 
 interface WorkflowRelayProps {
   visit: Visit;
